Support NSNumber conversions on 32-bit platforms
This fixes Int64 decoding on 32-bit platforms like the iPhone 4s.

The previous implementation copied memory from NSValue and this would
fail if you were going from Int32 -> Int64, because the memory would
only be partially initialized. The unit tests to catch this had been
commented out.

The new approach is higher level: it uses the NSNumber conversion methods
to solve this problem as well as providing the expected behavior in
overflow cases (such as converting a large Int64 to Int32)

This maintains backwards compatibility in case someone has added
NSValueCastable conformity to other types.